+++ plone/app/contenttypes/migration/dxmigration.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from plone.app.contenttypes.interfaces import IEvent
 from plone.app.contenttypes.migration.field_migrators import datetime_fixer
-from plone.app.contenttypes.migration.utils import HAS_MULTILINGUAL
 from plone.dexterity.interfaces import IDexterityContent
 from plone.dexterity.interfaces import IDexterityFTI
 from plone.event.utils import default_timezone
@@ -172,10 +171,7 @@ def migrate_base_class_to_new_class(obj,
 
 def list_of_objects_with_changed_base_class(context):
     catalog = getToolByName(context, 'portal_catalog')
-    query = {'object_provides': IDexterityContent.__identifier__}
-    if HAS_MULTILINGUAL and 'Language' in catalog.indexes():
-        query['Language'] = 'all'
-    for brain in catalog(query):
+    for brain in catalog(object_provides=IDexterityContent.__identifier__):
         try:
             obj = brain.getObject()
         except NotFound: