+++ plone/app/contenttypes/migration/utils.py
@@ -47,26 +47,9 @@ from zope.lifecycleevent import modified
 import json
 import logging
 import os
-import pkg_resources
 
 
 logger = logging.getLogger(__name__)
-
-# Is there a multilingual addon?
-try:
-    pkg_resources.get_distribution('Products.LinguaPlone')
-except pkg_resources.DistributionNotFound:
-    HAS_MULTILINGUAL = False
-else:
-    HAS_MULTILINGUAL = True
-
-if not HAS_MULTILINGUAL:
-    try:
-        pkg_resources.get_distribution('plone.app.multilingual')
-    except pkg_resources.DistributionNotFound:
-        HAS_MULTILINGUAL = False
-    else:
-        HAS_MULTILINGUAL = True
 
 
 def isSchemaExtended(iface):